pyproject.toml
[tool.poetry]
name = "scrape"
version = "0.1.0"
description = ""
authors = ["Julian Sotelo <juliansotelo@mail.fresnostate.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
ipykernel = "^6.29.5"
requests = "^2.32.3"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"
azure-storage-blob = "^12.22.0"
streamlit = "^1.38.0"
streamlit-pdf-viewer = "^0.0.17"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
